# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: World of Shadows — Backend API
  version: 1.0.0
  description: 'Machine-readable inventory of Flask `/api/v1` routes. Human-readable request/response examples: `docs/api/REFERENCE.md`.
    Tag taxonomy: `docs/api/openapi-taxonomy.md`.'
servers:
- url: /
  description: Same origin as backend (set base URL in your client).
...
paths:
...
  /api/v1/game/bootstrap:
    get:
      tags:
      - GameBootstrap
      summary: GET /game/bootstrap
      operationId: get_api_v1_game_bootstrap
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
  /api/v1/game/characters:
    get:
      tags:
      - GameBootstrap
      summary: GET /game/characters
      operationId: get_api_v1_game_characters
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
    post:
      tags:
      - GameBootstrap
      summary: POST /game/characters
      operationId: post_api_v1_game_characters
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
  /api/v1/game/characters/{character_id}:
    delete:
      tags:
      - GameBootstrap
      summary: DELETE /game/characters/{character_id}
      operationId: delete_api_v1_game_characters_character_id
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
      parameters:
      - &id013
        name: character_id
        in: path
        required: true
        schema:
          type: integer
    patch:
      tags:
      - GameBootstrap
      summary: PATCH /game/characters/{character_id}
      operationId: patch_api_v1_game_characters_character_id
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
      parameters:
      - *id013
  /api/v1/game/content/experiences:
    get:
      tags:
      - GameBootstrap
      summary: GET /game/content/experiences
      operationId: get_api_v1_game_content_experiences
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
    post:
      tags:
      - GameBootstrap
      summary: POST /game/content/experiences
      operationId: post_api_v1_game_content_experiences
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
  /api/v1/game/content/experiences/{experience_id}:
    get:
      tags:
      - GameBootstrap
      summary: GET /game/content/experiences/{experience_id}
      operationId: get_api_v1_game_content_experiences_experience_id
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
      parameters:
      - &id014
        name: experience_id
        in: path
        required: true
        schema:
          type: integer
    patch:
      tags:
      - GameBootstrap
      summary: PATCH /game/content/experiences/{experience_id}
      operationId: patch_api_v1_game_content_experiences_experience_id
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
      parameters:
      - *id014
  /api/v1/game/content/experiences/{experience_id}/governance/decision:
    post:
      tags:
      - GameBootstrap
      summary: POST /game/content/experiences/{experience_id}/governance/decision
      operationId: post_api_v1_game_content_experiences_experience_id_governance_decision
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
      parameters:
      - name: experience_id
        in: path
        required: true
        schema:
          type: integer
  /api/v1/game/content/experiences/{experience_id}/governance/mark-publishable:
    post:
      tags:
      - GameBootstrap
      summary: POST /game/content/experiences/{experience_id}/governance/mark-publishable
      operationId: post_api_v1_game_content_experiences_experience_id_governance_mark_publishable
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
      parameters:
      - name: experience_id
        in: path
        required: true
        schema:
          type: integer
  /api/v1/game/content/experiences/{experience_id}/governance/submit-review:
    post:
      tags:
      - GameBootstrap
      summary: POST /game/content/experiences/{experience_id}/governance/submit-review
      operationId: post_api_v1_game_content_experiences_experience_id_governance_submit_review
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
      parameters:
      - name: experience_id
        in: path
        required: true
        schema:
          type: integer
  /api/v1/game/content/experiences/{experience_id}/publish:
    post:
      tags:
      - GameBootstrap
      summary: POST /game/content/experiences/{experience_id}/publish
      operationId: post_api_v1_game_content_experiences_experience_id_publish
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
      parameters:
      - name: experience_id
        in: path
        required: true
        schema:
          type: integer
  /api/v1/game/content/experiences/{experience_id}/unpublish:
    post:
      tags:
      - GameBootstrap
      summary: POST /game/content/experiences/{experience_id}/unpublish
      operationId: post_api_v1_game_content_experiences_experience_id_unpublish
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
      parameters:
      - name: experience_id
        in: path
        required: true
        schema:
          type: integer
  /api/v1/game/content/published:
    get:
      tags:
      - GameBootstrap
      summary: GET /game/content/published
      operationId: get_api_v1_game_content_published
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Flask-JWT-Extended access token from `POST /api/v1/auth/login`.
security:
- BearerAuth: []